chore: move state types to correct file
